Add PMTiles support
It's now possible to output RGB tiles directly to PMTiles,
single file archive for tiled data just like MBTiles but
with better compression and possibly even speed.

The implementation copies the logic from mbtiler.py. It uses
a Writer from the pmtiles library which has been added as
a dependency.

There is a lot of code that is shared between these two and
it could definitely be improved a lot.